requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
prometheus-client==0.19.0
aiohttp==3.9.1
protobuf==4.25.1
